Combine related qstat tests in a single test suite
